+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Autocomplete JavaScript Node globals ([#85](https://github.com/anmonteiro/lumo/issues/85)).
 - Add `dir` REPL special ([#87](https://github.com/anmonteiro/lumo/issues/87)).
+- Add `-V / --version` CLI option to print the Lumo version.
 
 ### Changes